backend/requirements.txt
aioice==0.10.1
aiortc==1.13.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
av==14.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
google-crc32c==1.7.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.2
lap==0.5.12
mando==0.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
minio==7.2.15
mpmath==1.3.0
networkx==3.4.2
numpy
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.0
pydantic_core==2.33.0
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
radon==6.0.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.0
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.99
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1